Major update

+++ gkb-backend/Models/Stock.cs
@@ -7,7 +7,7 @@ namespace gkb_service.Models
         public int StockId { get; set; }
         public string? StockName { get; set; }
         public string? Unit { get; set; }
-        public int? UnitPrice { get; set; }
+        public float? UnitPrice { get; set; }
         public int? Availability { get; set; }
     }
 }
